The beginnings of a Temp Prob IO class. Some reorganization.
