+++ package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "body-parser": "^1.0.1",
     "debug": "^2.2.0",
+    "ds18x20": "^0.2.0",
     "express": "^4.0.0",
     "nconf": "^0.8.2"
   },
